+++ lib/aitch/html_parser.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-module Aitch
-  module HTMLParser
-    def self.load(source)
-      Nokogiri::HTML(source.to_s)
-    end
-  end
-end