+++ lib/aitch/response_parser/html_parser.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+module Aitch
+  module ResponseParser
+    module HTMLParser
+      def self.type
+        :html
+      end
+
+      def self.match?(content_type)
+        content_type =~ /html/
+      end
+
+      def self.load(source)
+        Nokogiri::HTML(source.to_s)
+      end
+    end
+  end
+end