+++ lib/aitch/xml_parser.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-module Aitch
-  module XMLParser
-    def self.load(source)
-      Nokogiri::XML(source.to_s, nil, "utf-8")
-    end
-  end
-end