+++ spec/aitch/response_parser/html_parser_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 require "spec_helper"
 
-describe Aitch::HTMLParser do
+describe Aitch::ResponseParser::HTMLParser do
   it "instantiates Nokogiri" do
     expect(Nokogiri).to receive(:HTML).with("HTML")
-    Aitch::HTMLParser.load("HTML")
+    Aitch::ResponseParser::HTMLParser.load("HTML")
   end
 end