+++ CHANGELOG.md
@@ -4,6 +4,7 @@
     * Use frozen string magic comments
     * 307 redirections will honor the HTTP request method
     * Fix issue with redirection; it was including the target url for one level redirection.
+    * Methods `Aitch::Response#html`, `Aitch::Response#json`, `Aitch::Response#xml` were removed; use `Aitch::Response#data` instead.
 * **0.5.0** - May 24, 2015
     * Remove hard dependency on ActiveSupport
     * Add response validation